fix(onnx): Phase-1 ships CPU/fp32; defer CoreML + ISA int8
CoreML EP crashes at inference on batch>1 (build_session guards session
construction only, not run()), and there is no universal int8 weight file
(published int8 variants are ISA-specific). Ship fp32/CPU on every reachable
host: universal, torch-free, and where 'never fatal' genuinely holds. Bind
probe made multi-text so a batch-breaking EP is rejected at bind (auto falls
through to ST), not mid-ingest. Validated end-to-end on the M4 Pro. CoreML,
OpenVINO, and ISA-aware int8 + an inference-time CPU fallback are a deferred
hardware-tuning unit.

+++ api/services/architecture.py
@@ -377,9 +377,11 @@ class UnknownArchitecture:
         return "5m"
 
     def recommended_onnx_providers(self) -> "OnnxExecutionPlan":
-        # Degraded mode: the universal floor, conservatively quantized.
+        # Degraded mode: the universal floor with the universal weight format.
+        # fp32 (onnx/model.onnx) exists for every model and runs on any CPU;
+        # ISA-specific int8 is deferred to the hardware-tuning phase.
         return OnnxExecutionPlan(
-            providers=["CPUExecutionProvider"], quant="int8", provider_options=[{}]
+            providers=["CPUExecutionProvider"], quant="fp32", provider_options=[{}]
         )
